refactor(storybook/website): rename stories
