+++ libs/website/shared/ui/navigation/index.ts
@@ -1,4 +1,4 @@
 export { FOOTER_CONSTANTS } from './footer/constants/footer.constants'
 export { FooterPresenter } from './footer/ui/footer.presenter'
 export { NAVBAR_CONSTANTS } from './navbar/constants/navbar.constants'
-export { NavbarPresenter } from './navbar/ui/navbar.presenter'
+export { NavbarContainer } from './navbar/ui/navbar.container'
